레벨 기반 구현

+++ second-poject/Assets/01. Scripts/Character/Character.cs
@@ -12,7 +12,7 @@ public class Character : MonoBehaviour
     public List<SO_Skill> skillList;
 
     public int Level = 1;
-    public float Exp = 0;
+    public int Exp = 0;
 
     protected virtual void Update()
     {
